requirements.txt
pandas
streamlit
python-dotenv
chromadb
sentence-transformers
langchain
langchain-groq
langchain-community
langchain-huggingface